requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.0.5
django-crispy-forms==1.10.0
django-widget-tweaks==1.4.8
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
rope==0.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
